requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
certifi==2025.11.12
click==8.3.1
dnspython==2.8.0
elastic-transport==9.2.1
elasticsearch==9.2.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
numpy==2.2.6
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
